requirements.txt.txt
fastapi
uvicorn
psycopg2-binary
python-jose
python-multipart
